Change data type passed to CThread::Join to long
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@26489 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/algo/blast/api/prelim_stage.cpp
@@ -212,7 +212,7 @@ CBlastPrelimSearch::x_LaunchMultiThreadedSearch()
     // ... and wait for the threads to finish
     bool error_occurred(false);
     NON_CONST_ITERATE(TBlastThreads, thread, the_threads) {
-        int result(0);
+        long result(0);
         (*thread)->Join(reinterpret_cast<void**>(&result));
         if (result != 0) {
             error_occurred = true;
